requirements.txt
hydra-core~=1.3.2
transformers~=4.37.2
omegaconf~=2.3.0
langchain-community~=0.2.1
langchain~=0.2.1
torch~=2.3.0
sentence-transformers~=2.7.0
wikipedia~=1.4.0
chromadb~=0.5.0